+++ src/com/ray/model/entities/User.java
@@ -6,12 +6,12 @@ public class User implements Serializable{
 
     private static final long serialVersionUID = 1L;
     
-    private Integer id;
+    private Long id;
     private String name;
     private String username;
     private String password;
     
-    public User(Integer id, String name, String username, String password) {
+    public User(Long id, String name, String username, String password) {
 	this.id = id;
 	this.name = name;
 	this.username = username;
@@ -21,11 +21,11 @@ public class User implements Serializable{
     public User() {
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
